separates bft tests into sub-packages and sub-ci jobs
