+++ .github/workflows/ci.yml
@@ -192,7 +192,9 @@ jobs:
       matrix:
         make:
           - make -C integration access-tests
-          - make -C integration bft-tests
+          - make -C integration bft-framework-tests
+          - make -C integration bft-protocol-tests
+          - make -C integration bft-gossipsub-tests
           - make -C integration collection-tests
           - make -C integration consensus-tests
           - make -C integration epochs-cohort1-tests
