+++ bors.toml
@@ -17,7 +17,9 @@ status = [
   "Unit Tests (utils)",
   "Unit Tests (others)",
   "Integration Tests (make -C integration access-tests)",
-  "Integration Tests (make -C integration bft-tests)",
+  "Integration Tests (make -C integration bft-framework-tests)",
+  "Integration Tests (make -C integration bft-gossipsub-tests)",
+  "Integration Tests (make -C integration bft-protocol-tests)",
   "Integration Tests (make -C integration collection-tests)",
   "Integration Tests (make -C integration consensus-tests)",
   "Integration Tests (make -C integration epochs-cohort1-tests)",
