+++ integration/tests/bft/framework/dummy_orchestrator.go
@@ -1,4 +1,4 @@
-package passthrough
+package framework
 
 import (
 	"errors"
