+++ integration/tests/bft/framework/passthrough_test.go
@@ -1,4 +1,4 @@
-package passthrough
+package framework
 
 import (
 	"testing"
