+++ integration/tests/bft/framework/suite.go
@@ -1,4 +1,4 @@
-package passthrough
+package framework
 
 import (
 	"github.com/rs/zerolog"
